requirements.txt
django==3.2.11
djangorestframework==3.13.1
drf-yasg==1.20.0
rest-framework-simplejwt==5.2.0
